GUI basic functions
Basic functions have been implemented for each element in the graphical interface, namely: join/quit/send buttons, list of connected users for each user and the text output used as server output log.
Now it can be simulated the joining of a user in the group. Also the list of users in the group is shown, which is updated every 2 seconds and is independent for each user. It is also possible to send messages, which will be displayed in the log section.